pyproject.toml
[project]
name = "mimo"
version = "0.1.0"
description = "Calculadora de Saldo a Favor - Declaración Anual"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jinja2>=3.1.0",
    "google-generativeai>=0.8.0",
    "packaging>=25.0",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
pdf = [
    "weasyprint>=62.0"
]
